backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.17
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
jinja2==3.1.4
httpx==0.28.1
typing_extensions>=4.12.0
